moved components to base folder. also moved css files to new css folder ( #14 )
