--- src/components/Notify/index.js
@@ -1,1 +1,0 @@
-export * from './Notify.js';
