+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import "antd/dist/antd.css";
 import './index.css';
-import { App } from './components/App/App'; 
+import { App } from './components/App'; 
 //import * as serviceWorker from './serviceWorker';
